refactor: Move loading data logic into page from TravelPlaceList
